fiddle w/ some build stuff.
