+++ apps/wallets/smart-wallet/next/package.json
@@ -9,7 +9,7 @@
         "start": "next start"
     },
     "dependencies": {
-        "@crossmint/client-sdk-react-ui": "workspace:*",
+        "@crossmint/client-sdk-react-ui": "1.3.7",
         "@radix-ui/react-avatar": "1.1.0",
         "@radix-ui/react-dropdown-menu": "2.1.1",
         "@radix-ui/react-slot": "1.1.0",
