+++ packages/client/auth-core/client/index.ts
@@ -1,2 +1,1 @@
-export { useAuth, AuthProvider, type AuthProviderParams } from "../src/contexts/authlogin";
-export { getCachedJwt } from "../src/utils/session";
+export { useAuth, AuthProvider, type AuthProviderParams, getCachedJwt } from "../src/contexts/authlogin";
