+++ packages/client/auth-core/src/contexts/authlogin.tsx
@@ -6,6 +6,14 @@ import { type ReactNode, createContext, useContext, useEffect, useMemo, useState
 
 import { Crossmint } from "@crossmint/common-sdk-base";
 
+export function getCachedJwt(): string | undefined {
+    if (typeof document === "undefined") {
+        return undefined; // Check if we're on the client-side
+    }
+    const crossmintSession = document.cookie.split("; ").find((row) => row.startsWith("crossmint-session"));
+    return crossmintSession ? crossmintSession.split("=")[1] : undefined;
+}
+
 type AuthContextType = {
     login: () => void;
     logout: () => void;
